sokol: complete toe's GL->sokol migration (engine builds + tests pass)
toe now renders entirely through sokol_gfx (native Metal on macOS); no GL
in the engine. All 7 tests pass.

  * App concept gains begin_frame(px, r,g,b) / end_frame(): the HOST owns the
    swapchain pass now (acquire drawable, begin pass + clear to the terminal's
    default bg, ... end pass, commit, present). toe draws INSIDE it. This is
    the clean sokol handoff \u2014 the host still owns the device/window, exactly
    like it owned the GL context before.\n  * run_loop: begin_frame -> session.render (+ overlay) -> end_frame -> present.
    Dropped the epoxy include, glViewport, glClear.
  * terminal.cpp: render()/render_overlay() no longer glBindFramebuffer \u2014 they
    just draw into the host's open pass. Added Session::default_bg() so the
    host knows the clear colour. Dropped epoxy.
  * renderer.hpp: default_bg() accessor.
  * CMake: build sokol_impl.m (Metal) + gpu.cpp, run sokol-shdc to regen the
    shader header, link Metal/QuartzCore frameworks on Apple; keep epoxy only
    on Linux (sokol GLCORE loader). Dropped the dead gfx/shader.cpp.
  * gpu.cpp: size_t fixes.

Deleted the old GL shader.cpp/shader.hpp (the Program wrapper). RenderContext
stays as a now-trivial capability token. Next: hand's CAMetalLayer backend
(begin_frame/end_frame + drawable) and removing epoxy from hand.

+++ src/gfx/gpu.cpp
@@ -5,6 +5,8 @@
 
 #include "toe/gfx/gpu.hpp"
 
+#include <cstddef>
+
 #include "sokol/sokol_gfx.h"
 
 namespace toe::gfx::gpu {
@@ -25,7 +27,7 @@ std::uint32_t make_image(int w, int h, Fmt fmt, const void *pixels) {
     sg_image img = sg_make_image(&d);
     if (img.id != SG_INVALID_ID && pixels) {
         sg_image_data data = {};
-        data.mip_levels[0] = {pixels, static_cast<std::size_t>(w) * h * bpp(fmt)};
+        data.mip_levels[0] = {pixels, static_cast<std::size_t>(w) * static_cast<std::size_t>(h) * static_cast<std::size_t>(bpp(fmt))};
         sg_update_image(img, &data);
     }
     return img.id;
@@ -35,7 +37,7 @@ void update_image(std::uint32_t image_id, int w, int h, Fmt fmt, const void *pix
     if (!image_id || !pixels) return;
     sg_image img = {image_id};
     sg_image_data data = {};
-    data.mip_levels[0] = {pixels, static_cast<std::size_t>(w) * h * bpp(fmt)};
+    data.mip_levels[0] = {pixels, static_cast<std::size_t>(w) * static_cast<std::size_t>(h) * static_cast<std::size_t>(bpp(fmt))};
     sg_update_image(img, &data);
 }
